WIP: `Catalog.vue` → `App.vue`
